refactor: make `getFormEmail` common
